Found out about the Random.Init(seed) function and realised that every seed related code i wrote was completely useless and way to complex. Changed and Simplified all the seed functions, since the Random Class is actually usable now.
Also added some new Scripts and a new Scene all called something with alternative;
Those are WiP:
I realised that my functionality may be more helpful if we just enable buildings in a certain area.
The Plan is that the User can dedicate certain areas using a cube Object where buildings can be generated.
When that is finished I will think about removing the original functionality and follow that one.